output generate tasks and presets on startup
